Sprite factory
I am testing the idea of having a factory that can generate an element
in the game from a javascript configuration object.
In this case the player is generated using a spriteFactory
(factory/sprite.factory.js) and the configuration object
(configs/player.js).
If the player is an object, a tool to generate this configuration file
can be created.